colorz/requirements.txt
aiohttp
Pillow
scipy
